+++ rust/src/audio/controls.rs
@@ -7,4 +7,4 @@ use crossbeam::channel::{Sender, Receiver};
 pub static TXRX:RwLock<Option<(Sender<bool>, Receiver<bool>)>> = RwLock::new(None);
 pub static IS_PLAYING:AtomicBool = AtomicBool::new(true);
 pub static VOLUME:RwLock<f32> = RwLock::new(1.0);
-pub static SEEK_TS:RwLock<Option<f64>> = RwLock::new(None);
+pub static SEEK_TS:RwLock<Option<u64>> = RwLock::new(None);
